[Input] Rename TextInput to Input
