+++ docs/data/api/input.json
@@ -1,0 +1,17 @@
+{
+  "props": {
+    "className": { "type": { "name": "union", "description": "func<br>&#124;&nbsp;string" } },
+    "render": { "type": { "name": "union", "description": "element<br>&#124;&nbsp;func" } }
+  },
+  "name": "Input",
+  "imports": ["import { Input } from '@base-ui-components/react/input';"],
+  "classes": [],
+  "spread": true,
+  "themeDefaultProps": true,
+  "muiName": "Input",
+  "forwardsRefTo": "HTMLInputElement",
+  "filename": "/packages/react/src/input/Input.tsx",
+  "inheritance": null,
+  "demos": "<ul><li><a href=\"/components/react-input/\">Text Input</a></li></ul>",
+  "cssComponent": false
+}
