+++ docs/data/components/form/FormIntroduction/system/index.js
@@ -4,7 +4,7 @@ import * as React from 'react';
 import { Form } from '@base-ui-components/react/form';
 import { Fieldset } from '@base-ui-components/react/fieldset';
 import { Field } from '@base-ui-components/react/field';
-import { TextInput } from '@base-ui-components/react/text-input';
+import { TextInput } from '@base-ui-components/react/input';
 import { styled } from '@mui/system';
 
 export default function FormIntroduction() {
