+++ docs/data/components/input/InputIntroduction/system/index.tsx.preview
@@ -1,0 +1,1 @@
+<Input />
