+++ docs/data/components/input/input.mdx
@@ -1,0 +1,28 @@
+---
+productId: base-ui
+title: React Input component
+description: Input is a UI element that lets users input single line text.
+components: Input
+githubLabel: 'component: input'
+packageName: '@base-ui-components/react'
+---
+
+# Text Input
+
+<Description />
+
+<ComponentLinkHeader design={false} />
+
+<Demo demo="InputIntroduction" defaultCodeOpen="false" bg="gradient" />
+
+## Installation
+
+<InstallationInstructions componentName="Input" />
+
+## Anatomy
+
+`<Input />` renders an `<input>`, enhanced with field context when placed inside a [`Field`](/components/react-field).
+
+```tsx
+<Input />
+```
