--- docs/data/components/text-input/TextInputIntroduction/system/index.tsx.preview
@@ -1,1 +1,0 @@
-<TextInput />
