--- docs/data/components/text-input/text-input.mdx
@@ -1,28 +1,0 @@
----
-productId: base-ui
-title: React TextInput component
-description: Text Input is a UI element that lets users input single line text.
-components: TextInput
-githubLabel: 'component: text-input'
-packageName: '@base-ui-components/react'
----
-
-# Text Input
-
-<Description />
-
-<ComponentLinkHeader design={false} />
-
-<Demo demo="TextInputIntroduction" defaultCodeOpen="false" bg="gradient" />
-
-## Installation
-
-<InstallationInstructions componentName="TextInput" />
-
-## Anatomy
-
-`<TextInput />` renders an `<input>`, enhanced with field context when placed inside a [`Field`](/components/react-field).
-
-```tsx
-<TextInput />
-```
