+++ packages/react/src/index.ts
@@ -20,5 +20,5 @@ export * from './separator';
 export * from './slider';
 export * from './switch';
 export * from './tabs';
-export * from './text-input';
+export * from './input';
 export * from './tooltip';
