--- packages/react/src/text-input/index.ts
@@ -1,1 +1,0 @@
-export { TextInput } from './TextInput';
